+++ NEWS.md
@@ -1,4 +1,6 @@
 # SIBER 2.1.2.9000
++ Contains a hotfix owing to a change to 'spatstat' detailed below
++ swapped package `spatstat` for `spatstat.utils` as per instructions from their package maintainers 23/03/2017
 
 # SIBER 2.1.2
 + added ability to specify custom pch point orders in `plotSiberObject()`